+++ test/dummy/app/controllers/dummy_portals_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class DummyPortalsController < ApplicationController
+  def show
+  end
+end